Add integration test
The approach taken to get this test to work is not elegant and could likely be improved. It adds an entirely new discovery type `srvtest` (for the sole purpose of testing, and admittedly polluting the main code), which simply inherits from `SrvUnicastHostsProvider` and overrides `buildResolver` to return the same hard-coded ports that `SrvDiscoveryIntegrationTest` uses when starting 2 nodes. But hey, it works!